Move track to its own service
